Move CA1001 to its new location - no functional change.
