--- PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/BasicTypesThatOwnDisposableFieldsShouldBeDisposable.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Microsoft.ApiDesignGuidelines.Analyzers     
-    ''' <summary>
-    ''' CA1001: Types that own disposable fields should be disposable
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicTypesThatOwnDisposableFieldsShouldBeDisposableFixer
-        Inherits TypesThatOwnDisposableFieldsShouldBeDisposableFixer 
-
-    End Class
-End Namespace
