+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.props
@@ -9,6 +9,7 @@
    <PropertyGroup>
       <CodeAnalysisRuleSetOverrides>
         $(CodeAnalysisRuleSetOverrides);
+        -Microsoft.Design#CA1001;
         -Microsoft.Design#CA1019;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
